+++ SH-Engine/src/shTime.c
@@ -5,8 +5,8 @@
 
 void shGetTime(ShTime* time) {
 	time->now = glfwGetTime();
-	time->delta_time = (float)time->now - time->last_time;
-	time->last_time = (float)time->now;
+	time->delta_time = time->now - time->last_time;
+	time->last_time = time->now;
 }
 
 void shSetTime(double now, ShTime *time) {
